% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/callapply.R
\name{callapply}
\alias{callapply}
\title{specifying function call to lapply output}
\usage{
callapply(func, x, FUN, ...)
}
\arguments{
\item{func}{A string for a function call to an `lapply` output. Common uses would be "rbind" or "cbind"}

\item{x}{A vector to iterate over for `lapply`}

\item{FUN}{A function supplied to `lapply`}

\item{...}{Other arguments passed to `lapply`}
}
\description{
Wrapper around `do.call` and `lapply`
}
