% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ind_fn.R
\name{ind_fn}
\alias{ind_fn}
\title{Check vectors are mutually in each other}
\usage{
ind_fn(terms_vec, vars)
}
\arguments{
\item{terms_vec}{vector of all terms to be split}

\item{vars}{a vector of variables of a term}
}
\description{
NA
}
