% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_conditions.R
\name{int_conditions}
\alias{int_conditions}
\title{int_conditions}
\usage{
int_conditions(mod, main_vars = NULL, demean = TRUE, data = NULL, names = NULL)
}
\arguments{
\item{mod}{A model object}

\item{main_vars}{A vector of variable names in the interaction of interest. If unspecified, takes by default variables from the highest order interaction}

\item{demean}{Should the function also return mean conditions?}

\item{data}{Dataset used when fitting model}

\item{names}{A named vector for renaming variables}
}
\value{
A data frame object that contains all effects, respective conditions, and estimated hypotheses
}
\description{
Conducts hypothesis testing across all conditions of interaction effects of a fitted model
}
\examples{
set.seed(1)
dat <- data.frame(X1 = sample(0:1, 100, replace=TRUE), X2 = sample(0:1, 100, replace=TRUE))
dat <- dat \%>\% mutate(Y = X1 + 2*X2 + 3*X1*X2)
mod <- lm(Y~X1*X2, dat)
cond_tab <- int_conditions(mod, main_vars = c("X1", "X2"), data = dat, names = c(A1 = "X1", A2 = "X2"))

}
