% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/int_graph.R
\name{int_graph}
\alias{int_graph}
\title{Graph conditional means and effects from a regression with interactions}
\usage{
int_graph(
  data,
  facet = NULL,
  digits = 3,
  width = 10,
  col_level = "gray95",
  col_effect = c(`0` = "dodgerblue", `1` = "chartreuse", all = "tomato", effect =
    "gold"),
  alpha = c(1, 0.5, 1, 0.5),
  col_label = "wheat"
)
}
\arguments{
\item{data}{output of a `int_conditions` call}

\item{facet}{A formula of the form X~Y with all interaction terms specified. Must be supplied.}

\item{digits}{How many decimal digits should be displayed?}

\item{width}{Maximum number of characters before wrapping the strip.}

\item{col_level}{Colour of background panel for all levels. Different colours should be used to differentiate between conditional levels and effects.}

\item{col_effect}{Colour of background panel for each effect type. Must be a named vector with names 0, 1, all, effect. Only needed if running 3 and 4 way interactions}

\item{alpha}{Alpha value for each facet row in a fourway interaction. The purpose of this is to create a dark, then light contrast between rows that are faceted.}

\item{col_label}{Background colour of label strip.}
}
\value{
A ggplot object that plots all conditional means and effects from the output of `int_conditions`
}
\description{
Wrapper around ggplot that graphs all conditional means and effects from the output of `int_conditions`
}
\examples{
set.seed(1)
dat <- data.frame(X1 = sample(0:1, 100, replace=TRUE), X2 = sample(0:1, 100, replace=TRUE))
dat <- dat |> mutate(Y = X1 + 2*X2 + 3*X1*X2 + rnorm(1))
mod <- lm(Y~X1*X2, dat)
cond_tab <- int_conditions(mod, data = dat, main_vars = c("X1", "X2"), names = c(A1 = "X1", A2 = "X2"))
plot <- int_graph(cond_tab, facet = X1~X2)

# Output is a ggplot object that can be manually manipulated further
plot + ggtitle("testing")

}
