% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_wrap_gen_both.R
\name{label_wrap_gen_both}
\alias{label_wrap_gen_both}
\title{Wrap labels}
\usage{
label_wrap_gen_both(width = 10, multi_line = TRUE, sep = ": ")
}
\arguments{
\item{width}{Maximum number of characters before wrapping the strip.}

\item{multi_line}{Whether to display the labels of multiple factors on separate lines.}

\item{sep}{String separating variables and values.}
}
\description{
Wrapper around ggplot's label_both function. Wraps and retains variable names in a facet label
}
