% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit_vec.R
\name{strsplit_vec}
\alias{strsplit_vec}
\title{unlist split strings}
\usage{
strsplit_vec(x, split, ...)
}
\arguments{
\item{x}{A string to be split.}

\item{split}{A regexp for the string to be split by.}

\item{...}{Additional arguments passed to `strsplit`}
}
\description{
Wrapper around `strsplit`
}
