% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toydata.R
\docType{data}
\name{toydata}
\alias{toydata}
\title{Toy data with interactions}
\format{
An object of class \code{"data.frame"}

#' \describe{
  \item{Col1}{description of Col1}
  \item{Col2}{description of Col2}
}
}
\usage{
data(toydata)
}
\description{
Small, artificially generated toy data set with interactions variables and fixed effect clusters
}
\references{
This data set was artificially created for the Rinteract package.
}
\keyword{datasets}
